Code refactoring. Fixed translations.
Moved inInRangeToRender3d from BytecodeHelper to RenderingOffsetter.
Modified set function in Config to use the local variables instead of reading from the config file to set the offsets value.

Added checks to make sure [::ofs set] and [::ofs reset] only work if they would result in a change to the player's offsets.
Added errors for when the previous occurs and players try to use non-numbers with the [::ofs set] command.