Refactoring continues. Found couple bugs with rendering and receiving player offsets on server, will fix in later commits.
Bug 1: Rendered model is short from physical offset position. This inaccuracy increases with distance.
Bug 2: If other players are not nearby the client when logging in their current offsets will not be received. And, if they change theirs while the client is on, and they are far away in non-loaded client chunks, the same thing will occur.